Raster Splitter Rename files and algo
