+++ enmapbox/apps/SpecDeepMap/__init__.py
@@ -2,7 +2,7 @@ from enmapbox.gui.applications import EnMAPBoxApplication
 
 try:
     from enmapbox.apps.SpecDeepMap.processingalgorithmDS_SUM12 import DatasetSplitter_SUM
-    from enmapbox.apps.SpecDeepMap.processingalgorithm_Raster_Splitter import RasterSplitter
+    from enmapbox.apps.SpecDeepMap.processingalgorithm_raster_splitter import RasterSplitter
     from enmapbox.apps.SpecDeepMap.processingalgorithm_DL_UNET50_MOD_15_059_16 import DL_Train_MOD
     from enmapbox.apps.SpecDeepMap.processingalgorithm_PRED_GT_NO_DATA_mod11 import DL_Mapper
     from enmapbox.apps.SpecDeepMap.processingalgorithm_Tensorboard2 import Tensorboard
